refactor(backup): use OptionSheet, extract BackupSheet, move logic to BackupService, remove auto-backup UI
refactor(backup): update German strings to match English changes
