+++ app/src/main/java/de/pawcode/cardstore/data/services/BackupService.kt
@@ -11,6 +11,7 @@ import de.pawcode.cardstore.data.models.BackupCardData
 import de.pawcode.cardstore.data.models.BackupData
 import de.pawcode.cardstore.data.models.BackupLabelData
 import java.time.Instant
+import kotlinx.coroutines.flow.first
 import kotlinx.serialization.json.Json
 
 object BackupService {
@@ -66,6 +67,28 @@ object BackupService {
         }
     }
     return true
+  }
+
+  suspend fun performCreateBackup(
+    context: Context,
+    uri: Uri,
+    cardRepository: de.pawcode.cardstore.data.database.repositories.CardRepository,
+    labelRepository: de.pawcode.cardstore.data.database.repositories.LabelRepository,
+  ): Result<Unit> = runCatching {
+    val cards = cardRepository.allCards.first()
+    val labels = labelRepository.allLabels.first()
+    createBackup(context, uri, cards, labels)
+  }
+
+  suspend fun performRestoreBackup(
+    context: Context,
+    uri: Uri,
+    cardRepository: de.pawcode.cardstore.data.database.repositories.CardRepository,
+  ): Result<Unit> = runCatching {
+    val data = readBackup(context, uri)
+    if (!validateBackup(data)) error("Invalid backup")
+    val (cards, labels, crossRefs) = toEntities(data)
+    cardRepository.restoreBackup(cards, labels, crossRefs)
   }
 
   fun toEntities(
