+++ app/src/main/java/de/pawcode/cardstore/ui/screens/AboutScreen.kt
@@ -1,15 +1,10 @@
 package de.pawcode.cardstore.ui.screens
 
-import android.app.backup.BackupManager
 import android.content.Intent
 import android.content.pm.PackageInfo
-import android.net.Uri
 import android.widget.Toast
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -19,17 +14,9 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -46,18 +33,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import de.pawcode.cardstore.R
-import de.pawcode.cardstore.data.database.repositories.CardRepository
-import de.pawcode.cardstore.data.database.repositories.LabelRepository
 import de.pawcode.cardstore.data.managers.PreferencesManager
-import de.pawcode.cardstore.data.services.BackupService
 import de.pawcode.cardstore.data.services.BiometricAuthService
 import de.pawcode.cardstore.navigation.Navigator
 import de.pawcode.cardstore.ui.components.AppBar
 import de.pawcode.cardstore.ui.components.SettingsGroup
 import de.pawcode.cardstore.ui.components.SettingsItem
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-import kotlinx.coroutines.flow.first
+import de.pawcode.cardstore.ui.sheets.BackupSheet
 import kotlinx.coroutines.launch
 
 data class Technology(val name: String, val url: String, @param:DrawableRes val icon: Int)
@@ -111,7 +93,6 @@ val TECHNOLOGIES =
     ),
   )
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AboutScreen(navigator: Navigator) {
   val context = LocalContext.current
@@ -126,156 +107,14 @@ fun AboutScreen(navigator: Navigator) {
     e.printStackTrace()
   }
 
-  val isAutoBackupEnabled = runCatching {
-    val bm = BackupManager(context)
-    val method = bm.javaClass.getMethod("isBackupEnabled")
-    method.invoke(bm) as Boolean
-  }
-    .getOrDefault(false)
-
-  val cardRepository = CardRepository(context)
-  val labelRepository = LabelRepository(context)
-
   var showBackupSheet by remember { mutableStateOf(false) }
-  var pendingRestoreUri by remember { mutableStateOf<Uri?>(null) }
-  var showRestoreConfirmDialog by remember { mutableStateOf(false) }
 
-  val createBackupLauncher =
-    rememberLauncherForActivityResult(ActivityResultContracts.CreateDocument("application/json")) {
-      uri ->
-      if (uri == null) return@rememberLauncherForActivityResult
-      scope.launch {
-        runCatching {
-          val cards = cardRepository.allCards.first()
-          val labels = labelRepository.allLabels.first()
-          BackupService.createBackup(context, uri, cards, labels)
-          Toast.makeText(context, context.getString(R.string.backup_success), Toast.LENGTH_SHORT)
-            .show()
-        }
-          .onFailure {
-            Toast.makeText(context, context.getString(R.string.backup_failed), Toast.LENGTH_SHORT)
-              .show()
-          }
-      }
-    }
-
-  val restoreBackupLauncher =
-    rememberLauncherForActivityResult(ActivityResultContracts.OpenDocument()) { uri ->
-      if (uri == null) return@rememberLauncherForActivityResult
-      pendingRestoreUri = uri
-      showRestoreConfirmDialog = true
-    }
-
-  fun triggerCreateBackup() {
-    if (biometricEnabled) {
-      val activity = context as? androidx.fragment.app.FragmentActivity ?: return
-      BiometricAuthService.authenticate(
-        activity = activity,
-        title = context.getString(R.string.backup_biometric_title),
-        subtitle = context.getString(R.string.backup_biometric_subtitle),
-        onSuccess = {
-          val date = LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE)
-          createBackupLauncher.launch("cardstore-backup-$date.json")
-        },
-        onError = {},
-      )
-    } else {
-      val date = LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE)
-      createBackupLauncher.launch("cardstore-backup-$date.json")
-    }
-  }
-
-  if (showRestoreConfirmDialog) {
-    AlertDialog(
-      onDismissRequest = {
-        showRestoreConfirmDialog = false
-        pendingRestoreUri = null
-      },
-      title = { Text(stringResource(R.string.restore_confirm_title)) },
-      text = { Text(stringResource(R.string.restore_confirm_body)) },
-      confirmButton = {
-        TextButton(
-          onClick = {
-            showRestoreConfirmDialog = false
-            val uri = pendingRestoreUri ?: return@TextButton
-            pendingRestoreUri = null
-            scope.launch {
-              runCatching {
-                val data = BackupService.readBackup(context, uri)
-                if (!BackupService.validateBackup(data)) error("Invalid backup")
-                val (cards, labels, crossRefs) = BackupService.toEntities(data)
-                cardRepository.restoreBackup(cards, labels, crossRefs)
-                Toast.makeText(
-                    context,
-                    context.getString(R.string.restore_success),
-                    Toast.LENGTH_SHORT,
-                  )
-                  .show()
-              }
-                .onFailure {
-                  Toast.makeText(
-                      context,
-                      context.getString(R.string.restore_failed),
-                      Toast.LENGTH_SHORT,
-                    )
-                    .show()
-                }
-            }
-          }
-        ) {
-          Text(stringResource(android.R.string.ok))
-        }
-      },
-      dismissButton = {
-        TextButton(
-          onClick = {
-            showRestoreConfirmDialog = false
-            pendingRestoreUri = null
-          }
-        ) {
-          Text(stringResource(android.R.string.cancel))
-        }
-      },
-    )
-  }
-
-  if (showBackupSheet) {
-    ModalBottomSheet(
-      onDismissRequest = { showBackupSheet = false },
-      sheetState = rememberModalBottomSheetState(),
-    ) {
-      Column(modifier = Modifier.fillMaxWidth().padding(bottom = 32.dp)) {
-        ListItem(
-          modifier =
-            Modifier.fillMaxWidth().clickable {
-              showBackupSheet = false
-              triggerCreateBackup()
-            },
-          headlineContent = { Text(stringResource(R.string.backup_create)) },
-          leadingContent = {
-            Icon(painter = painterResource(R.drawable.save_solid), contentDescription = null)
-          },
-        )
-        ListItem(
-          modifier =
-            Modifier.fillMaxWidth().clickable {
-              showBackupSheet = false
-              restoreBackupLauncher.launch(arrayOf("application/json"))
-            },
-          headlineContent = { Text(stringResource(R.string.backup_restore)) },
-          leadingContent = {
-            Icon(painter = painterResource(R.drawable.file_open_solid), contentDescription = null)
-          },
-        )
-      }
-    }
-  }
+  BackupSheet(visible = showBackupSheet, onDismiss = { showBackupSheet = false })
 
   AboutScreenComponent(
     packageInfo = packageInfo,
     biometricAvailable = BiometricAuthService.isBiometricAvailable(context),
     biometricEnabled = biometricEnabled,
-    isAutoBackupEnabled = isAutoBackupEnabled,
     onBack = { navigator.goBack() },
     onOpenWebsite = { context.startActivity(Intent(Intent.ACTION_VIEW, it.toUri())) },
     onBiometricToggle = { enabled ->
@@ -315,7 +154,6 @@ fun AboutScreenComponent(
   packageInfo: PackageInfo,
   biometricAvailable: Boolean,
   biometricEnabled: Boolean,
-  isAutoBackupEnabled: Boolean,
   onBack: () -> Unit,
   onOpenWebsite: (String) -> Unit,
   onBiometricToggle: (Boolean) -> Unit,
@@ -376,21 +214,6 @@ fun AboutScreenComponent(
                 enabled = biometricAvailable,
               )
             },
-          )
-
-          SettingsItem(
-            icon =
-              painterResource(
-                if (isAutoBackupEnabled) R.drawable.cloud_done_solid else R.drawable.cloud_off_solid
-              ),
-            iconColor = MaterialTheme.colorScheme.onSecondaryFixedVariant,
-            iconBackground = MaterialTheme.colorScheme.primaryFixed,
-            title = stringResource(R.string.auto_backup_title),
-            subtitle =
-              stringResource(
-                if (isAutoBackupEnabled) R.string.auto_backup_active
-                else R.string.auto_backup_inactive
-              ),
           )
 
           SettingsItem(
@@ -483,7 +306,6 @@ fun PreviewAboutScreenComponent() {
     packageInfo = packageInfo,
     biometricAvailable = true,
     biometricEnabled = true,
-    isAutoBackupEnabled = true,
     onBack = {},
     onOpenWebsite = {},
     onBiometricToggle = {},
